Add a task to fill the constituency party column
Since the constituency by postcode Parliament API doesn't return
party we have to have a separate task to query a different API.
This API returns a list of members which by default returns all
members ever but can be filtered down to just active members.

If we don't find the member id in the list then we assume that
they've been marked as inactive for some reasons - either a
general election or by-election due to resignation/death.